Refactor import statement for contributors in cards.jsx
